# docs/swagger.yaml
definitions:
  domain.Event:
    properties:
      capacity:
        type: integer
      category_id:
        type: integer
      creation_time:
        type: string
      description:
        type: string
      ending_time:
        type: string
      event_uuid:
        type: string
      image_url:
        type: string
      location:
        $ref: '#/definitions/domain.Location'
      organizer_id:
        type: integer
      organizer_name:
        type: string
      starting_time:
        type: string
      status:
        type: string
      tags:
        items:
          type: string
        type: array
      ticketsId:
        items:
          $ref: '#/definitions/domain.Ticket'
        type: array
      title:
        type: string
    type: object
  domain.Location:
    properties:
      address:
        description: '"Dostyk Ave 85, Almaty, Kazakhstan"'
        type: string
      event_uuid.omitempty:
        type: string
      latitude:
        description: "43.2405"
        type: number
      location_id:
        type: integer
      longitude:
        description: "76.9312"
        type: number
      name:
        description: '"Dostyk Plaza"'
        type: string
    type: object
  domain.Ticket:
    properties:
      currency:
        type: string
      event_uuid:
        type: string
      name:
        type: string
      price:
        type: number
      quantity:
        type: integer
      sold:
        type: integer
      ticket_uuid:
        type: string
    type: object
...
info:
  contact: {}
paths:
...
  /api/event:
    delete:
      parameters:
      - description: Event UUID
        in: query
        name: event_uuid
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Delete event
      tags:
      - events
    get:
      parameters:
      - description: Event UUID
        in: query
        name: event_uuid
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get event
      tags:
      - events
    post:
      consumes:
      - application/json
      description: Creates an event and assigns the authenticated user as its organizer.
      parameters:
      - description: Event to create
        in: body
        name: event
        required: true
        schema:
          $ref: '#/definitions/domain.Event'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Create event
      tags:
      - events
  /api/event/all:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: List events
      tags:
      - events
  /api/event/category:
    get:
      parameters:
      - description: Category ID
        in: query
        name: category_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: List events by category
      tags:
      - events
  /api/event/top-ten:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get trending events
      tags:
      - events
  /api/event/update/image:
    post:
      parameters:
      - description: Event UUID
        in: query
        name: event_uuid
        required: true
        type: string
      - description: Image URL
        in: query
        name: image_url
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Update event image URL
      tags:
      - events
  /api/event/update/upload:
    post:
      consumes:
      - multipart/form-data
      parameters:
      - description: Event UUID
        in: query
        name: event_uuid
        required: true
        type: string
      - description: Image file (PNG, JPEG, or WebP; up to 10 MiB)
        in: formData
        name: image
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Upload event image
      tags:
      - events
  /api/images/{image}:
    get:
      parameters:
      - description: Image filename
        in: path
        name: image
        required: true
        type: string
      produces:
      - image/png
      - image/jpeg
      - image/webp
      responses:
        "200":
          description: OK
          schema:
            type: file
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get stored image
      tags:
      - images
...
swagger: "2.0"